+++ src/OFS/interfaces.py
@@ -566,8 +566,7 @@ class IObjectManager(IZopeObject, ICopyContainer, INavigation, IManageable,
         """Return a list of subobjects, used by tree tag.
         """
 
-    def manage_exportObject(id='', download=None, toxml=None,
-                            RESPONSE=None,REQUEST=None):
+    def manage_exportObject(id='', download=None, RESPONSE=None, REQUEST=None):
         """Exports an object to a file and returns that file."""
 
     def manage_importObject(file, REQUEST=None, set_owner=1):